requirements.txt
streamlit>=1.28.0
langgraph>=0.0.40
langchain>=0.1.0
langchain-deepseek>=0.0.1
langchain-community>=0.0.20
newsapi-python>=0.1.6
python-dotenv>=1.0.0
pydantic>=2.5.0
pandas>=2.1.0
openai>=1.6.0
anthropic>=0.18.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pyyaml>=6.0.1
markdown>=3.5.0
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0

